The function pay seat and food for customer

+++ src/main/java/org/group3/project_swp391_bookingmovieticket/repositories/ISeatRepository.java
@@ -10,5 +10,5 @@ import java.util.List;
 public interface ISeatRepository extends JpaRepository<Seat, Integer> {
 
     List<Seat> findByRoomId(int roomId);
-    List<Seat> findByIdIn(List<Long> ids);
+    List<Seat> findByIdIn(List<Integer> ids);
 }
